% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractAnalyze.R
\name{abstractAnalyze}
\alias{abstractAnalyze}
\title{abstractAnalyze}
\arguments{
\item{variants}{[[character]] A character vector containing the rs numbers of SNPs.}

\item{directory}{[[character]] A character string containing a path to the directory where the SNP data is stored.}

\item{tissues}{[[character]] A character vector containing the tissues of interest.}

\item{clustering}{[[character]] A character string indicating which clustring method should be used: `agnes` or `PAM`.}
}
\value{
[[list]]
}
\description{
abstractAnalyze
}
\keyword{internal}
