% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnesPearson.R
\name{agnesPearson}
\alias{agnesPearson}
\title{agnesPearson}
\arguments{
\item{x}{[[matrix]] A matrix containg the gene expression.}

\item{k}{[[numeric]] A integer number indicating the number of clusters.}
}
\value{
[[list]]
}
\description{
agnesPearson
}
\keyword{internal}
