% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseSNPs.R
\name{AnalyseSNPs}
\alias{AnalyseSNPs}
\title{AnalyseSNPs}
\usage{
AnalyseSNPs(eQTLData, tissue, clustering)
}
\arguments{
\item{eQTLData}{[[data.frame]] A data.frame containing the eQTL data.}

\item{tissue}{[[character]] A character vector containing the tissues of interest.}

\item{clustering}{[[character]] A character string indicating which clustring method should be used: `agnes` or `PAM`}
}
\value{
[[list]]
}
\description{
AnalyseSNPs
}
\keyword{internal}
