% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOdds.R
\name{calculateOdds}
\alias{calculateOdds}
\title{calculateOdds}
\arguments{
\item{pathway}{[[character]] A character vector containing the genes that are included in a pathway.}

\item{geneSet}{[[character]] A character vector containing the found genes.}

\item{background}{[[character]] A character vector containing a set of background genes.}
}
\value{
[[vector]]
}
\description{
calculateOdds
}
\keyword{internal}
