% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromInteractionsToBED.R
\name{chromInteractionsToBED}
\alias{chromInteractionsToBED}
\title{chromInteractionsToBED}
\arguments{
\item{interactions}{[[GRanges]] A Genomic ranges object containing the chromatin interactions.}

\item{RangeOfInterest}{[[]]}
}
\value{
[[list]]
}
\description{
chromInteractionsToBED
}
\keyword{internal}
