% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquer-package.R
\docType{package}
\name{conquer-package}
\alias{conquer-package}
\title{conquer: Convolution-type smoothed quantile regression}
\description{
Fast and accurate conquer (convolution-type smoothed quantile regression) implemented with Barzilai-Borwein gradient descent with a Huber warm start.
Statistical inference for conquer via multiplier bootstrap is also included.
}
\references{
Barzilai, J. and Borwein, J. M. (1988). Two-point step size gradient methods. IMA J. Numer. Anal. 8 141–148.

Fernandes, M., Guerre, E. and Horta, E. (2019). Smoothing quantile regressions. J. Bus. Econ. Statist., in press.

Koenker, R. (2005). Quantile Regression. Cambridge University Press, Cambridge.

Koenker, R. Package "quantreg".

Koenker, R. and Bassett, G. (1978). Regression quantiles. Econometrica 46 33-50.
}
\author{
Xiaoou Pan <xip024@ucsd.edu>, Kean Ming Tan <keanming@umich.edu>, Wen-Xin Zhou <wez243@ucsd.edu>
}
