% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smqr.R
\name{conquer}
\alias{conquer}
\title{Convolution-type smoothed quantile regression}
\usage{
conquer(
  X,
  Y,
  tau = 0.5,
  kernel = c("Gaussian", "uniform", "parabolic", "triangular"),
  h = 0,
  standardize = TRUE,
  tol = 1e-05,
  iteMax = 5000,
  ci = FALSE,
  alpha = 0.05,
  B = 1000
)
}
\arguments{
\item{X}{The design matrix with dimension \eqn{n} by \eqn{p}, where \eqn{p < n}.}

\item{Y}{The response vector with length \eqn{n}.}

\item{tau}{(\strong{optional}) The desired quantile level of the regression problem. The value must be in \eqn{(0, 1)}.}

\item{kernel}{(\strong{optional}) A character string specifying the kernel function. Choices include "Gaussian" (default), "uniform", "parabolic" or "triangular".}

\item{h}{(\strong{optional}) The bandwidth of kernel smoothing. The value will be \eqn{((log(n) + p) / n)^0.4} without specific input or if the input value \eqn{< 0.05}.}

\item{standardize}{(\strong{optional}) A logical flag. If \code{standardize = TRUE}, then the design matrix will be standardized so that each column has \eqn{0} mean and unit standard deviation.}

\item{tol}{(\strong{optional}) Tolerance level of gradient descent.}

\item{iteMax}{(\strong{optional}) Maximum of iteration for gradient descent.}

\item{ci}{(\strong{optional}) A logical flag. If \code{ci = TRUE}, then three types of confidence intervals (percentile, pivotal and normal) will be constructed via multiplier bootstrap.}

\item{alpha}{(\strong{optional}) The nominal noncoverage probability for the confidence intervals. The value must be in \eqn{(0, 1)}.}

\item{B}{(\strong{optional}) The size of bootstrap sample.}
}
\value{
An object containing the following items will be returned:
\itemize{
\item \code{coeff} The estimated coefficients including the intercept. A \eqn{(p + 1)}-vector.
\item \code{ite} The total number of iteration of the gradient descent algorithm.
\item \code{residual} The residuals of the fitted quantile regression.
\item \code{bandwidth} The bandwidth value.
\item \code{tau} The desired quantile level of the regression problem.
\item \code{kernel} The choice of kernel function.
\item \code{n} The sample size.
\item \code{p} The dimension.
\item \code{perCI} The percentile confidence intervals for regression coefficients. Not available if \code{ci = FALSE}
\item \code{pivCI} The pivotal confidence intervals for regression coefficients. Not available if \code{ci = FALSE}
\item \code{normCI} The normal-based confidence intervals for regression coefficients. Not available if \code{ci = FALSE}
}
}
\description{
Smoothed quantile regression with fast computation and accurate estimation.
}
\examples{
n = 500; p = 10
beta = rep(1, p)
X = matrix(rnorm(n * p), n, p)
Y = 1 + X \%*\% beta + rt(n, 2)

## Smoothed quantile regression with Gaussian kernel
fit.Gauss = conquer(X, Y, tau = 0.5, kernel = "Gaussian")
beta.hat.Gauss = fit.Gauss$coeff

## Smoothe quantile regression with uniform kernel
fit.unif = conquer(X, Y, tau = 0.5, kernel = "uniform")
beta.hat.unif = fit.unif$coeff

## Construct three types of confidence intervals via multiplier bootstrap
fit = conquer(X, Y, tau = 0.5, kernel = "Gaussian", ci = TRUE)
ci.per = fit$perCI
ci.piv = fit$pivCI
ci.norm = fit$normCI
}
\references{
Barzilai, J. and Borwein, J. M. (1988). Two-point step size gradient methods. IMA J. Numer. Anal. 8 141–148.

Fernandes, M., Guerre, E. and Horta, E. (2019). Smoothing quantile regressions. J. Bus. Econ. Statist., in press.

Koenker, R. and Bassett, G. (1978). Regression quantiles. Econometrica 46 33-50.
}
