% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.pathways.R
\name{extract.pathways}
\alias{extract.pathways}
\title{Extract enriched pathways}
\usage{
extract.pathways(SNPSummary, subset = NULL)
}
\value{
A `dataframe` of enriched pathways
}
\description{
Extract enriched pathways
}
