% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameToVersionedID.R
\name{GeneNameToVersionedID}
\alias{GeneNameToVersionedID}
\title{GeneNameToVersionedID}
\arguments{
\item{geneNames}{}
}
\value{
[[data.frame]]
}
\description{
GeneNameToVersionedID
}
\keyword{internal}
