% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eQTL_precalculated.R
\name{get_eQTL_precalculated}
\alias{get_eQTL_precalculated}
\title{get_eQTL_bulk}
\usage{
get_eQTL_precalculated(lead, tissues)
}
\value{
[[data.frame]]
}
\description{
get_eQTL_bulk
}
\keyword{internal}
