% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromatinInteractionData.R
\name{getChromatinInteractionData}
\alias{getChromatinInteractionData}
\title{getChromatinInteractionData}
\arguments{
\item{topHits}{}
}
\value{
[[IRanges]]
}
\description{
getChromatinInteractionData
}
\keyword{internal}
