% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoexpression.R
\name{getCoexpression}
\alias{getCoexpression}
\title{getCoexpression}
\arguments{
\item{geneSet}{Vector of genes}

\item{expressionData}{data.frame with expression data with samples as rows and genes as columns}

\item{method}{Which correlation coefficient to calculate "pearson", "spearman" or "kendall"}

\item{threshold}{Threshold of minimum correlation coefficient on which to base the co-expression}
}
\value{
[[vector]]
}
\description{
getCoexpression
}
\keyword{internal}
