% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichREnrichment.R
\name{getEnrichREnrichment}
\alias{getEnrichREnrichment}
\title{importFrom httr POST
import enrichR
import ggplot2}
\usage{
getEnrichREnrichment(tissue, SNPSummary)
}
\description{
importFrom httr POST
import enrichR
import ggplot2
}
