% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geteQTLdata.R
\name{geteQTLdata}
\alias{geteQTLdata}
\title{geteQTLdata}
\usage{
geteQTLdata(lead, Genes, parallel)
}
\arguments{
\item{lead}{leadSNP}

\item{Genes}{Gene names}

\item{parallel}{Parallel disabled}
}
\value{
[[data.frame]]
}
\description{
geteQTLdata
}
