% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantInfoEnsembl.R
\name{getVariantInfoEnsembl}
\alias{getVariantInfoEnsembl}
\title{getVariantInfoEnsembl}
\arguments{
\item{population}{}
}
\value{
[[list]]
}
\description{
getVariantInfoEnsembl
}
\keyword{internal}
