% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_KEGG.R
\name{prepare_KEGG}
\alias{prepare_KEGG}
\title{This function was obtained from clusterProfiler and was not exported. Therefore this function has been included here}
\usage{
prepare_KEGG(species, KEGG_Type = "KEGG", keyType = "kegg")
}
\arguments{
\item{species}{[[character]]}

\item{KEGG_Type}{[[character]]}

\item{keyType}{[[character]]}
}
\description{
This function was obtained from clusterProfiler and was not exported. Therefore this function has been included here
}
