% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\docType{data}
\name{resources}
\alias{resources}
\title{Dataframe of the data used in CONQUER
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)}
\format{
Datafram
}
\usage{
resources
}
\description{
Dataframe of the data used in CONQUER
GTEx Analysis V8 (dbGaP Accession phs000424.v8.p2)
}
\keyword{datasets}
