% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Collect data for one or more SNPs and perform an optional integrative analysis on multiple tissues.}
\usage{
summarize(
  variants,
  precalculated = TRUE,
  multiAnalyze = FALSE,
  tissues = NULL,
  directory = NULL,
  token = NULL,
  population = "CEU"
)
}
\arguments{
\item{variants}{vector Vector with one or more SNP names (rs*)}

\item{precalculated}{Default is TRUE. GTEx by default only takes along SNPs 1MB from the
start site. If TRUE, only the precalculated SNPs by GTEx are used. When FALSE, CONQUER will
calculate all pairwise comparisons which may take substantially longer (up to days for large number of SNPs).
Note that this may yield more interesting results as GTEx may miss interesting QTLs.}

\item{multiAnalyze}{Default is FALSE. If TRUE, an integrated analysis will be performed across the SNPs.}

\item{tissues}{Tissues of interest. Should be a vector of names based on teh}

\item{directory}{String of the directory where the files should be stored for the SNPs investigated}

\item{token}{Token required for LDlink, which can be obtained from the LDlink website}

\item{population}{Letter code of population of interest for example " CEU"}
}
\description{
Collect data for one or more SNPs and perform an optional integrative analysis on multiple tissues.
}
\examples{
\dontrun{
library(CONQUER)
summarize(variants = c("rs1558902","rs601945"),
directory=getwd(),
multiAnalyze=FALSE,
token="sometoken",
tissues=NULL)}
}
