% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDashboard.R
\name{visualizeDashboard}
\alias{visualizeDashboard}
\title{visualizeDashboard}
\usage{
visualizeDashboard(loadedSNPs, SNPSummary, ColocSummary)
}
\value{
[[NULL]]
}
\description{
visualizeDashboard
}
\keyword{internal}
