% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAgnesSpearman.R
\name{ClusterAgnesSpearman}
\alias{ClusterAgnesSpearman}
\title{Cluster Agnes /Spearman}
\usage{
ClusterAgnesSpearman(expressionData)
}
\value{
[[data.frame]]
}
\description{
Cluster Agnes /Spearman
}
