% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquer-package.R
\docType{package}
\name{conquer-package}
\alias{conquer-package}
\title{Conquer: Convolution-Type Smoothed Quantile Regression}
\description{
Fast and accurate convolution-type smoothed quantile regression for low/high-dimensional estimation and inference.
}
\references{
Barzilai, J. and Borwein, J. M. (1988). Two-point step size gradient methods. IMA J. Numer. Anal. 8 141–148.

Belloni, A. and Chernozhukov, V. (2011). l_1 penalized quantile regression in high-dimensional sparse models. Ann. Statist. 39 82-130.

Fan, J., Liu, H., Sun, Q. and Zhang, T. (2018). I-LAMM for sparse learning: Simultaneous control of algorithmic complexity and statistical error. Ann. Statist. 46 814-841.

Fernandes, M., Guerre, E. and Horta, E. (2019). Smoothing quantile regressions. J. Bus. Econ. Statist. 39 338-357.

He, X., Pan, X., Tan, K. M., and Zhou, W.-X. (2021+). Smoothed quantile regression for large-scale inference. J. Econometrics, in press.

Horowitz, J. L. (1998). Bootstrap methods for median regression models. Econometrica 66 1327–1351.

Koenker, R. (2005). Quantile Regression. Cambridge University Press, Cambridge.

Koenker, R. Package "quantreg".

Koenker, R. and Bassett, G. (1978). Regression quantiles. Econometrica 46 33-50.

Pan, X. and Zhou, W.-X. (2020). Multiplier bootstrap for quantile regression: Non-asymptotic theory under random design. Information and Inference: A Journal of the IMA 10 813-861.
}
\author{
Xuming He <xmhe@umich.edu>, Xiaoou Pan <xip024@ucsd.edu>, Kean Ming Tan <keanming@umich.edu>, and Wen-Xin Zhou <wez243@ucsd.edu>
}
