% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareChromatinStates.R
\name{PrepareChromatinStates}
\alias{PrepareChromatinStates}
\title{PrepareChromatinStates}
\usage{
PrepareChromatinStates(SNPData, sampleSel = NULL)
}
\arguments{
\item{SNPData}{SNPData}

\item{sampleSel}{sampleSel}
}
\value{
[[list]]
}
\description{
PrepareChromatinStates
}
