% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Collect data for one or more SNPs and perform an optional integrative analysis on multiple tissues.}
\usage{
summarize(
  variants,
  multiAnalyze = FALSE,
  tissues = NULL,
  directory = NULL,
  token = NULL,
  population = "CEU"
)
}
\arguments{
\item{variants}{vector Vector with one or more SNP names (rs*)}

\item{multiAnalyze}{Default is FALSE. If TRUE, an integrated analysis will be performed across the SNPs.}

\item{tissues}{Tissues of interest. Should be a vector of names based on teh}

\item{directory}{String of the directory where the files should be stored for the SNPs investigated}

\item{token}{Token required for LDlink, which can be obtained from the LDlink website}

\item{population}{Letter code of population of interest for example " CEU"}
}
\description{
Collect data for one or more SNPs and perform an optional integrative analysis on multiple tissues.
}
\examples{
\dontrun{
library(CONQUER)
summarize(variants = c("rs1558902","rs601945"),
directory=getwd(),
multiAnalyze=FALSE,
token="sometoken",
tissues=NULL)}
}
