% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Main function to visualize the investigated SNPs}
\usage{
visualize(directory, SNPs)
}
\arguments{
\item{directory}{[[character]] Character string of the directory in which the SNP data and optional summary files are stored}

\item{SNPs}{[[vector]] Vector of SNP names that correspond to the SNP names provided in summarize. Can also be a single SNP name, but should always be a rs* number.}
}
\description{
Main function to visualize the investigated SNPs
}
\examples{
\dontrun{summarize("somedirectory","rs1558902")}
}
